Add and use our own EDN-parsing middleware.
'ring.middleware.edn/wrap-edn-params unfortunately requires the 'request :body
to be a map, but om-sync POSTs whatever collection is at the cursor in a :create
transaction, and it may not be a map.

This new middleware is less robust than 'ring.middleware.edn/wrap-edn-params,
but is good enough for the purposes of the tutorial.